work

+++ 00●cloud_analy.py
@@ -9,32 +9,19 @@ from file_process import files_list, move_document
 from read_ht_excel_cloud0712 import convert_xlsb
 from read_ht_excel_cloud0712 import read_ctc_ht_excel
 from read_tc_excel_cloud0712 import read_tc_excel
-from text_gen import file_path_process
+from text_gen import copy_plan_file
 from text_gen import text_gen
-from text_gen import copy_stand_file
 
 file_doc_num = ""
 file_source_folder = "/workspaces/Auto-Work-Station/00source"
 file_dest_folder = "/workspaces/Auto-Work-Station/00dest"
 test_folder = "/workspaces/Auto-Work-Station/08Reference_Files"
 
-# 測試file_path_process、move_docutment
-# 輸入基本訊息，並顯示各項目路徑
-# Done
-
-
-def test_folder_path():
-    x = [file_doc_num, file_source_folder,
-         file_dest_folder] = file_path_process()
-    for asss in x:
-        print(asss)
-    # move_docutment(file_source_folder, file_dest_folder)
-
-# 測試雲端解析興達文件 Done 2023/07/20
-# 1.複製資料夾  2.分析Excel檔案  3.生成套印文字
-
-
 def test_cloud_ht():
+    '''
+    測試雲端解析興達文件 Done 2023/07/20
+    1.複製資料夾  2.分析Excel檔案  3.生成套印文字
+    '''
     sample_ht_folder = "/workspaces/Auto-Work-Station/08Reference_Files/HT"
     destination_dir = "/workspaces/Auto-Work-Station/00dest"
     # 複製Sample資料夾結構
@@ -64,8 +51,8 @@ def test_cloud_ht():
 # 測試雲端解析台中文件 Done 2023/07/20
 # 1.複製資料夾  2.分析Excel檔案  3.生成套印文字
 
-
 def test_cloud_tc():
+    '''測試雲端台中檔案'''
     sample_tc_folder = "/workspaces/Auto-Work-Station/08Reference_Files/TC"
     destination_dir = "/workspaces/Auto-Work-Station/00dest"
     # 複製Sample資料夾結構
@@ -87,35 +74,10 @@ def test_cloud_tc():
                  letter_num_value, letter_date_value, the_file)
         input("enter any keys to exit")
 
-# 測試路徑是否存在 Done 2023/07/20
-
-
-def test_path():
-    file_source_folder = "/workspaces/Auto-Work-Station/00source"
-    file_dest_folder = "/workspaces/Auto-Work-Station/00dest"
-    paths = [file_source_folder, file_dest_folder]
-    for path in paths:
-        if os.path.exists(path):
-            print(f"路徑 {path} 確實存在")
-        else:
-            print(f"路徑 {path} 不存在")
-
-# 測試生成套印文字
-
-
-def test_text_gen():
-    dest_folder = r"/workspaces/Auto-Work-Station/00source"
-    text_gen('HRSG Chimney-General Arrangement of Concrete Roof & Layout of Permanent Shutter to Roof Slab',
-             'A', 'TPC-TC(C0)-CD-23-0002', '2023/02/02', dest_folder)
-    input("Press enter to exit...")
-    return None
-
-# 總流程測試
-
-
 def cloud_total_run():
-    ROOT_GIT = r"/workspaces/Auto-Work-Station"
-    sys.path.append(ROOT_GIT)
+    """總流程測試"""
+    root_git = r"/workspaces/Auto-Work-Station"
+    sys.path.append(root_git)
 
     sample_folder = "/workspaces/Auto-Work-Station/00source"
     destination_dir = "/workspaces/Auto-Work-Station/00dest"
@@ -152,7 +114,7 @@ def cloud_total_run():
                 text_gen(letter_titl_value, drawing_vision_value,
                          letter_num_value, letter_date_value, destination_dir)
                 # 複製套印、傳真檔案
-                copy_stand_file(the_file)
+                copy_plan_file(the_file)
                 input("enter any keys to exit")
         else:
             pass
@@ -173,16 +135,15 @@ def cloud_total_run():
                 text_gen(letter_title, letter_vision, letter_num,
                          letter_date_value, destination_dir)
                 # 複製套印、傳真檔案
-                copy_stand_file(the_file)
+                copy_plan_file(the_file)
                 input("enter any keys to exit")
         input("----END----")
-
 
 if __name__ == '__main__':
     file_dest_folder = "/workspaces/Auto-Work-Station/00dest"
     # test_folder_path()
-
-    shutil.rmtree(file_dest_folder)
+    if os.path.exists(file_dest_folder):
+        shutil.rmtree(file_dest_folder)
     # test_cloud_ht()
 
     # shutil.rmtree(file_dest_folder)
